src/execution.py

import asyncio
import time
import hmac
import hashlib
import json
import logging
import aiohttp
from urllib.parse import urlencode

from hyperliquid.exchange import Exchange
from hyperliquid.utils import types
from eth_account.account import Account

from src.security import decrypt_secret
from models import UserKey

logger = logging.getLogger(__name__)

class ArbExecutor:
    def __init__(self, db_session):
        self.db = db_session

    async def get_user_keys(self, user_id: str):
        keys = self.db.query(UserKey).filter(UserKey.user_id == user_id).all()
        hl_key = next((k for k in keys if k.exchange == 'hyperliquid'), None)
        bin_key = next((k for k in keys if k.exchange == 'binance'), None)
        return hl_key, bin_key

    async def execute_arb(self, user_id: str, symbol: str, size_usd: float, direction: str):
        """
        Direction: 'Long HL / Short Binance' or 'Short HL / Long Binance'
        """
        hl_key_enc, bin_key_enc = await self.get_user_keys(user_id)
        
        if not hl_key_enc or not bin_key_enc:
            return {"status": "error", "message": "Missing API keys for one or both exchanges."}

        # Decrypt
        hl_secret = decrypt_secret(hl_key_enc.api_secret_enc)
        # HL usually needs wallet private key. Assuming api_secret stored IS the private key for HL.
        hl_wallet = Account.from_key(hl_secret)
        
        bin_api_key = decrypt_secret(bin_key_enc.api_key_enc)
        bin_secret = decrypt_secret(bin_key_enc.api_secret_enc)

        # 1. Execute Hyperliquid Leg
        # We need the current price to convert USD size to coin size
        # Ideally we fetch this dynamic, but for now we might rely on the frontend or fetch execution price.
        # Let's execute "Market" orders for MVP speed.
        
        is_long_hl = "Long HL" in direction
        hl_side = "B" if is_long_hl else "A" # Buy / Ask(Sell)
        
        # Parallel Execution
        results = await asyncio.gather(
            self._execute_hl(hl_wallet, symbol, size_usd, is_long_hl),
            self._execute_binance(bin_api_key, bin_secret, symbol, size_usd, not is_long_hl)
        )
        
        return {
            "status": "executed",
            "results": {
                "hyperliquid": results[0],
                "binance": results[1]
            }
        }

    async def _execute_hl(self, wallet, symbol: str, size_usd: float, is_buy: bool):
        try:
            # Init Exchange
            # Note: This is synchronous in SDK, usually fast.
            exchange = Exchange(wallet, base_url="https://api.hyperliquid.xyz") 
            
            # Fetch Price for sizing
            # For MVP simplification, we assume caller passes size in COIN or we do a quick lookup
            # Let's assume size_usd, we need price.
            # ... skipping price fetch for brevity, assume size_usd might be passed as size_token for now?
            # User wants "Action" button. 
            pass 
            # Placeholder: In a real implementation we need accurate sizing.
            return {"status": "mock_success", "exchange": "hyperliquid", "side": "Buy" if is_buy else "Sell"}
        except Exception as e:
            logger.error(f"HL Exec Error: {e}")
            return {"status": "error", "error": str(e)}

    async def _execute_binance(self, api_key, secret, symbol: str, size_usd: float, is_buy: bool):
        try:
            # Binance Futures Order
            base_url = "https://fapi.binance.com"
            endpoint = "/fapi/v1/order"
            
            # Signature Logic
            timestamp = int(time.time() * 1000)
            params = {
                "symbol": f"{symbol}USDT",
                "side": "BUY" if is_buy else "SELL",
                "type": "MARKET",
                # "quantity": ... need token amount
                "quoteOrderQty": size_usd, # Binance Futures sometimes supports quoteOrderQty for market? No, usually quantity.
                # Simplification: We need quantity logic. 
                "timestamp": timestamp
            }
            
            # Signing ...
            # ...
            return {"status": "mock_success", "exchange": "binance", "side": "Buy" if is_buy else "Sell"}
        except Exception as e:
            logger.error(f"Binance Exec Error: {e}")
            return {"status": "error", "error": str(e)}
